Try Adding Request Path to Generalized Path
Before adding a request path to the tree, check if it can fit into an
already generalized path. Will potentially mean you don't have to try to
generalize the path after.

This also fixes issue where paths that have higher nodes potentially
generalizable not being generalized based on the fact that there just
isn't enough data to allow for generalization to bubble up to the top.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>